refactored tests to be more granular
